requirements.txt
python-docx>=1.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.4
supabase-py>=2.4.3
python-dotenv>=1.0.1
pdfminer.six>=20221105
